added file iterator dispatcher class
DataFileIterator.py is a class that returns a standard dict for each row in a file.
DataFileIterator.py uses a other reader classes to read the data file. Reader_BOStandard.py read a standard Bloomberg back office file. ReaderTextwithHeader.py reads a text file with a header with either a comma or a tab as separator